vsock: refactor TSI proxy variants into multi-platform structures
This change refactors the TSI proxy components (`tsi_stream`, `tsi_dgram`, and `unix_proxy`) inside the `vsock` module. It extracts the files into cross-platform directory layouts (`mod.rs` and `unix.rs`), paving the way for Windows (WHP) compatibility by isolating POSIX/Unix socket logic. Additionally, socket errors in the `ProxyError` enumeration are changed to use standard library `std::io::Error` handles rather than `nix::errno::Errno`.
