+++ src/devices/src/virtio/vsock/tsi_stream/mod.rs
@@ -1,0 +1,381 @@
+use std::collections::HashMap;
+use std::num::Wrapping;
+use std::path::PathBuf;
+use std::sync::{Arc, Mutex};
+
+use super::super::Queue as VirtQueue;
+use super::defs;
+use super::defs::uapi;
+use super::muxer::{MuxerRx, push_packet};
+use super::muxer_rxq::MuxerRxQ;
+use super::packet::{TsiAcceptReq, TsiConnectReq, TsiListenReq, TsiSendtoAddr, VsockPacket};
+
+#[cfg(unix)]
+use nix::sys::socket::AddressFamily;
+#[cfg(unix)]
+use std::os::fd::OwnedFd;
+#[cfg(unix)]
+use std::os::unix::io::{AsRawFd, RawFd};
+
+use super::proxy::{Proxy, ProxyError, ProxyRemoval, ProxyStatus, ProxyUpdate, RecvPkt};
+use utils::epoll::EventSet;
+use vm_memory::GuestMemoryMmap;
+
+#[cfg(unix)]
+mod unix;
+#[cfg(unix)]
+use unix as sys;
+
+pub struct TsiStreamProxy {
+    pub(crate) id: u64,
+    pub(crate) cid: u64,
+    pub(crate) parent_id: u64,
+    pub(crate) family: AddressFamily,
+    pub(crate) local_port: u32,
+    pub(crate) peer_port: u32,
+    pub(crate) control_port: u32,
+    pub(crate) fd: OwnedFd,
+    pub status: ProxyStatus,
+    pub(crate) mem: GuestMemoryMmap,
+    pub(crate) queue: Arc<Mutex<VirtQueue>>,
+    pub(crate) rxq: Arc<Mutex<MuxerRxQ>>,
+    pub(crate) rx_cnt: Wrapping<u32>,
+    pub(crate) tx_cnt: Wrapping<u32>,
+    pub(crate) last_tx_cnt_sent: Wrapping<u32>,
+    pub(crate) peer_buf_alloc: u32,
+    pub(crate) peer_fwd_cnt: Wrapping<u32>,
+    pub(crate) push_cnt: Wrapping<u32>,
+    pub(crate) pending_accepts: u64,
+    pub(crate) unixsock_path: Option<PathBuf>,
+}
+
+impl TsiStreamProxy {
+    #[allow(clippy::too_many_arguments)]
+    pub fn new(
+        id: u64,
+        cid: u64,
+        family: u16,
+        local_port: u32,
+        peer_port: u32,
+        control_port: u32,
+        mem: GuestMemoryMmap,
+        queue: Arc<Mutex<VirtQueue>>,
+        rxq: Arc<Mutex<MuxerRxQ>>,
+    ) -> Result<Self, ProxyError> {
+        let (fd, family) = sys::create_socket(id, family)?;
+
+        Ok(TsiStreamProxy {
+            id,
+            cid,
+            parent_id: 0,
+            family,
+            local_port,
+            peer_port,
+            control_port,
+            fd,
+            status: ProxyStatus::Idle,
+            mem,
+            queue,
+            rxq,
+            rx_cnt: Wrapping(0),
+            tx_cnt: Wrapping(0),
+            last_tx_cnt_sent: Wrapping(0),
+            peer_buf_alloc: 0,
+            peer_fwd_cnt: Wrapping(0),
+            push_cnt: Wrapping(0),
+            pending_accepts: 0,
+            unixsock_path: None,
+        })
+    }
+
+    #[allow(clippy::too_many_arguments)]
+    pub fn new_reverse(
+        id: u64,
+        cid: u64,
+        parent_id: u64,
+        family: AddressFamily,
+        local_port: u32,
+        peer_port: u32,
+        fd: OwnedFd,
+        mem: GuestMemoryMmap,
+        queue: Arc<Mutex<VirtQueue>>,
+        rxq: Arc<Mutex<MuxerRxQ>>,
+    ) -> Self {
+        debug!("new_reverse: id={id} local_port={local_port} peer_port={peer_port}");
+        TsiStreamProxy {
+            id,
+            cid,
+            parent_id,
+            family,
+            local_port,
+            peer_port,
+            control_port: 0,
+            fd,
+            status: ProxyStatus::ReverseInit,
+            mem,
+            queue,
+            rxq,
+            rx_cnt: Wrapping(0),
+            tx_cnt: Wrapping(0),
+            last_tx_cnt_sent: Wrapping(0),
+            peer_buf_alloc: 0,
+            peer_fwd_cnt: Wrapping(0),
+            push_cnt: Wrapping(0),
+            pending_accepts: 0,
+            unixsock_path: None,
+        }
+    }
+
+    pub(crate) fn init_data_pkt(&self, pkt: &mut VsockPacket) {
+        debug!(
+            "init_data_pkt: id={}, local_port={}, peer_port={}",
+            self.id, self.local_port, self.peer_port
+        );
+        pkt.set_op(uapi::VSOCK_OP_RW)
+            .set_src_cid(uapi::VSOCK_HOST_CID)
+            .set_dst_cid(self.cid)
+            .set_src_port(self.local_port)
+            .set_dst_port(self.peer_port)
+            .set_type(uapi::VSOCK_TYPE_STREAM)
+            .set_buf_alloc(defs::CONN_TX_BUF_SIZE as u32)
+            .set_fwd_cnt(self.tx_cnt.0);
+    }
+
+    pub(crate) fn peer_avail_credit(&self) -> usize {
+        (Wrapping(self.peer_buf_alloc) - (self.rx_cnt - self.peer_fwd_cnt)).0 as usize
+    }
+
+    pub(crate) fn recv_pkt(&mut self) -> (bool, bool) {
+        let mut have_used = false;
+        let mut wait_credit = false;
+        let mut queue = self.queue.lock().unwrap();
+
+        while let Some(head) = queue.pop(&self.mem) {
+            let len = match VsockPacket::from_rx_virtq_head(&head) {
+                Ok(mut pkt) => match sys::recv_to_pkt(self, &mut pkt) {
+                    RecvPkt::WaitForCredit => {
+                        wait_credit = true;
+                        0
+                    }
+                    RecvPkt::Read(cnt) => {
+                        self.rx_cnt += Wrapping(cnt as u32);
+                        self.init_data_pkt(&mut pkt);
+                        pkt.set_len(cnt as u32);
+                        pkt.hdr().len() + cnt
+                    }
+                    RecvPkt::Close => {
+                        self.status = ProxyStatus::Closed;
+                        0
+                    }
+                    RecvPkt::Error => 0,
+                },
+                Err(e) => {
+                    debug!("recv_pkt: RX queue error: {e:?}");
+                    0
+                }
+            };
+
+            if len == 0 {
+                queue.undo_pop();
+                break;
+            } else {
+                have_used = true;
+                self.push_cnt += Wrapping(len as u32);
+                debug!(
+                    "recv_pkt: pushing packet with {} bytes, push_cnt={}",
+                    len, self.push_cnt
+                );
+                if let Err(e) = queue.add_used(&self.mem, head.index, len as u32) {
+                    error!("failed to add used elements to the queue: {e:?}");
+                }
+            }
+        }
+
+        debug!("recv_pkt: have_used={have_used}");
+        (have_used, wait_credit)
+    }
+
+    pub(crate) fn push_connect_rsp(&self, result: i32) {
+        debug!(
+            "push_connect_rsp: id: {}, control_port: {}, result: {}",
+            self.id, self.control_port, result
+        );
+
+        // This response goes to the control port (DGRAM).
+        let rx = MuxerRx::ConnResponse {
+            local_port: 1025,
+            peer_port: self.control_port,
+            result,
+        };
+        push_packet(self.cid, rx, &self.rxq, &self.queue, &self.mem);
+    }
+
+    pub(crate) fn push_reset(&self) {
+        debug!(
+            "push_reset: id: {}, peer_port: {}, local_port: {}",
+            self.id, self.peer_port, self.local_port
+        );
+
+        // This response goes to the connection.
+        let rx = MuxerRx::Reset {
+            local_port: self.local_port,
+            peer_port: self.peer_port,
+        };
+        push_packet(self.cid, rx, &self.rxq, &self.queue, &self.mem);
+    }
+}
+
+impl Proxy for TsiStreamProxy {
+    fn id(&self) -> u64 {
+        self.id
+    }
+
+    fn status(&self) -> ProxyStatus {
+        self.status
+    }
+
+    fn connect(&mut self, pkt: &VsockPacket, req: TsiConnectReq) -> ProxyUpdate {
+        sys::do_connect(self, pkt, req)
+    }
+
+    fn confirm_connect(&mut self, pkt: &VsockPacket) -> Option<ProxyUpdate> {
+        debug!(
+            "confirm_connect: local_port={} peer_port={}, src_port={}, dst_port={}",
+            pkt.dst_port(),
+            pkt.src_port(),
+            self.local_port,
+            self.peer_port,
+        );
+
+        self.peer_buf_alloc = pkt.buf_alloc();
+        self.peer_fwd_cnt = Wrapping(pkt.fwd_cnt());
+
+        self.local_port = pkt.dst_port();
+        self.peer_port = pkt.src_port();
+
+        // This response goes to the connection.
+        let rx = MuxerRx::OpResponse {
+            local_port: pkt.dst_port(),
+            peer_port: pkt.src_port(),
+        };
+        push_packet(self.cid, rx, &self.rxq, &self.queue, &self.mem);
+
+        // Now that the vsock transport is fully established, start listening
+        // for events in the TCP socket again.
+        Some(ProxyUpdate {
+            polling: Some((self.id, self.fd.as_raw_fd(), EventSet::IN)),
+            ..Default::default()
+        })
+    }
+
+    fn getpeername(&mut self, pkt: &VsockPacket) {
+        sys::do_getpeername(self, pkt)
+    }
+
+    fn sendmsg(&mut self, pkt: &VsockPacket) -> ProxyUpdate {
+        sys::sendmsg(self, pkt)
+    }
+
+    fn sendto_addr(&mut self, _req: TsiSendtoAddr) -> ProxyUpdate {
+        ProxyUpdate::default()
+    }
+
+    fn listen(
+        &mut self,
+        pkt: &VsockPacket,
+        req: TsiListenReq,
+        host_port_map: &Option<HashMap<u16, u16>>,
+    ) -> ProxyUpdate {
+        sys::do_listen(self, pkt, req, host_port_map)
+    }
+
+    fn accept(&mut self, req: TsiAcceptReq) -> ProxyUpdate {
+        sys::do_accept(self, req)
+    }
+
+    fn update_peer_credit(&mut self, pkt: &VsockPacket) -> ProxyUpdate {
+        sys::update_peer_credit(self, pkt)
+    }
+
+    fn push_op_request(&self) {
+        debug!(
+            "push_op_request: id={}, local_port={} peer_port={}",
+            self.id, self.local_port, self.peer_port
+        );
+
+        // This packet goes to the connection.
+        let rx = MuxerRx::OpRequest {
+            local_port: self.local_port,
+            peer_port: self.peer_port,
+        };
+        push_packet(self.cid, rx, &self.rxq, &self.queue, &self.mem);
+    }
+
+    fn process_op_response(&mut self, pkt: &VsockPacket) -> ProxyUpdate {
+        sys::process_op_response(self, pkt)
+    }
+
+    fn enqueue_accept(&mut self) {
+        debug!("enqueue_accept: control_port: {}", self.control_port);
+
+        if self.status == ProxyStatus::WaitingOnAccept {
+            self.status = ProxyStatus::Listening;
+            self.push_accept_rsp(0);
+        } else {
+            self.pending_accepts += 1;
+        }
+    }
+
+    fn push_accept_rsp(&self, result: i32) {
+        debug!(
+            "push_accept_rsp: control_port: {}, result: {}",
+            self.control_port, result
+        );
+
+        // This packet goes to the control port (DGRAM).
+        let rx = MuxerRx::AcceptResponse {
+            local_port: 1030,
+            peer_port: self.control_port,
+            result,
+        };
+        push_packet(self.cid, rx, &self.rxq, &self.queue, &self.mem);
+    }
+
+    fn shutdown(&mut self, pkt: &VsockPacket) {
+        sys::do_shutdown(self, pkt)
+    }
+
+    fn release(&mut self) -> ProxyUpdate {
+        debug!(
+            "release: id={}, tx_cnt={}, last_tx_cnt={}",
+            self.id, self.tx_cnt, self.last_tx_cnt_sent
+        );
+        let remove_proxy = if self.status == ProxyStatus::Listening {
+            ProxyRemoval::Immediate
+        } else {
+            ProxyRemoval::Deferred
+        };
+        ProxyUpdate {
+            remove_proxy,
+            ..Default::default()
+        }
+    }
+
+    fn process_event(&mut self, evset: EventSet) -> ProxyUpdate {
+        sys::process_event(self, evset)
+    }
+}
+
+impl AsRawFd for TsiStreamProxy {
+    fn as_raw_fd(&self) -> RawFd {
+        self.fd.as_raw_fd()
+    }
+}
+
+impl Drop for TsiStreamProxy {
+    fn drop(&mut self) {
+        if let Some(path) = &self.unixsock_path {
+            _ = std::fs::remove_file(path);
+        }
+    }
+}
